feat: JogoServiceImpl
